Move shard file to separated package and split its contents into multiple files. Added comments and docs. ShardManager and Shard constructors are now private. [ci skip]
