+++ nyxx/lib/src/internal/shard/Shard.dart
@@ -1,63 +1,5 @@
 part of nyxx;
 
-class ShardManager implements Disposable {
-  /// Emitted when the shard is ready.
-  late Stream<Shard> onConnected = this._onConnect.stream;
-
-  /// Emitted when the shard encounters an error.
-  late Stream<Shard> onDisconnect = this._onDisconnect.stream;
-
-  /// Emitted when shard receives member chunk.
-  late Stream<MemberChunkEvent> onMemberChunk = this._onMemberChunk.stream;
-
-  final StreamController<Shard> _onConnect = StreamController<Shard>.broadcast();
-  final StreamController<Shard> _onDisconnect = StreamController<Shard>.broadcast();
-  final StreamController<MemberChunkEvent> _onMemberChunk = StreamController.broadcast();
-
-  final Logger _logger = Logger("Shard Manager");
-
-  final _WS _ws;
-  final int _numShards;
-  final Map<int, Shard> _shards = {};
-
-  /// List of shards
-  Iterable<Shard> get shards => List.unmodifiable(_shards.values);
-
-  /// Starts shard manager
-  ShardManager(this._ws, this._numShards) {
-    _connect(_numShards - 1);
-  }
-
-  /// Sets presences on every shard
-  void setPresence(PresenceBuilder presenceBuilder) {
-    for (final shard in shards) {
-      shard.setPresence(presenceBuilder);
-    }
-  }
-
-  void _connect(int shardId) {
-    if(shardId < 0) {
-      return;
-    }
-
-    final shard = Shard(shardId, this, _ws.gateway);
-    _shards[shardId] = shard;
-
-    Future.delayed(const Duration(seconds: 1, milliseconds: 500), () => _connect(shardId - 1));
-  }
-
-  @override
-  Future<void> dispose() async {
-    for(final shard in this._shards.values) {
-      await shard.dispose();
-    }
-
-    await this._onConnect.close();
-    await this._onDisconnect.close();
-    await this._onMemberChunk.close();
-  }
-}
-
 class Shard implements Disposable {
   /// Id of shard
   final int id;
@@ -72,42 +14,44 @@ class Shard implements Disposable {
   ///
   /// To calculate the gateway latency, nyxx measures the time it takes for Discord to answer the gateway
   /// heartbeat packet with a heartbeat ack packet. Note this value is updated each time gateway responses to ack.
-  Duration get gatewayLatency => _gatewaylatency;
+  Duration get gatewayLatency => _gatewayLatency;
+
+  /// Returns true if shard is connected to websocket
+  bool get connected => _connected;
 
   late final Isolate _shardIsolate; // Reference to isolate
   late final Stream<dynamic> _receiveStream; // Broadcast stream on which data from isolate is received
   late final ReceivePort _receivePort; // Port on which data from isolate is received
   late final SendPort _isolateSendPort; // Port on which data can be sent to isolate
-  String? _sessionId;
-  int _sequence = 0;
-  late Timer _heartbeatTimer;
-  bool connected = false;
-  bool resume = false;
+  late SendPort _sendPort; // Sendport for isolate
 
-  late SendPort sendPort;
+  String? _sessionId; // Id of gateway session
+  int _sequence = 0; // Event sequence
+  late Timer _heartbeatTimer; // Heartbeat time
+  bool _connected = false; // Connection status
+  bool _resume = false; // Resume status
 
-  Duration _gatewaylatency = Duration();
-  late DateTime _lastHeartbeatSent;
-  bool _heartbeatAckReceived = false;
-
-  /// Isolate
-  Shard(this.id, this.manager, String gatewayUrl) {
+  Duration _gatewayLatency = const Duration(); // latency of discord
+  late DateTime _lastHeartbeatSent; // Datetime when last heartbeat was sent
+  bool _heartbeatAckReceived = false; // True if last heartbeat was acked
+  
+  Shard._new(this.id, this.manager, String gatewayUrl) {
     this._receivePort = ReceivePort();
     this._receiveStream = _receivePort.asBroadcastStream();
     this._isolateSendPort = _receivePort.sendPort;
 
-    Isolate.spawn(_shardHandler, _isolateSendPort).then((value) async {
-      this._shardIsolate = value;
-      this.sendPort = await _receiveStream.first as SendPort;
+    Isolate.spawn(_shardHandler, _isolateSendPort).then((isolate) async {
+      this._shardIsolate = isolate;
+      this._sendPort = await _receiveStream.first as SendPort;
 
-      this.sendPort.send({"cmd" : "INIT", "gatewayUrl" : gatewayUrl });
+      this._sendPort.send({"cmd" : "INIT", "gatewayUrl" : gatewayUrl });
       this._receiveStream.listen(_handle);
     });
   }
 
   /// Sends WS data.
   void send(int opCode, dynamic d) {
-    this.sendPort.send({"cmd": "SEND", "data" : {"op": opCode, "d": d}});
+    this._sendPort.send({"cmd": "SEND", "data" : {"op": opCode, "d": d}});
   }
 
   /// Updates clients voice state for [Guild] with given [guildId]
@@ -181,6 +125,7 @@ class Shard implements Disposable {
   void _handleError(dynamic data) {
     final closeCode = data["errorCode"] as int;
 
+    this._connected = false;
     this._heartbeatTimer.cancel();
     manager._logger.severe(
         "Shard $id disconnected. Error code: [${data['errorCode']}] | Error message: [${data['errorReason']}]");
@@ -214,55 +159,56 @@ class Shard implements Disposable {
   // Connects to gateway
   void _connect() {
 
-    this.resume = false;
-    Future.delayed(const Duration(seconds: 2), () => this.sendPort.send({ "cmd" : "CONNECT"}))
+    this._resume = false;
+    Future.delayed(const Duration(seconds: 2), () => this._sendPort.send({ "cmd" : "CONNECT"}))
         .then((_) => manager._logger.info("Connecting to gateway!"));
   }
 
   // Reconnects to gateway
   void _reconnect() {
-    this.resume = true;
-    Future.delayed(const Duration(seconds: 1), () => this.sendPort.send({ "cmd" : "CONNECT"}))
+    this._resume = true;
+    Future.delayed(const Duration(seconds: 1), () => this._sendPort.send({ "cmd" : "CONNECT"}))
         .then((value) => manager._logger.info("Resuming connection to gateway!"));
   }
 
-  Future<void> _handle(dynamic data) async {
-    if(data["cmd"] == "CONNECT_ACK") {
+  Future<void> _handle(dynamic rawData) async {
+    if(rawData["cmd"] == "CONNECT_ACK") {
       manager._logger.info("Shard $id connected to gateway!");
 
       return;
     }
 
-    if(data["cmd"] == "ERROR" || data["cmd"] == "DISCONNECTED") {
-      _handleError(data);
-    }
-
-    if(data["jsonData"] == null) {
+    if(rawData["cmd"] == "ERROR" || rawData["cmd"] == "DISCONNECTED") {
+      _handleError(rawData);
       return;
     }
 
-    final msg = data["jsonData"] as Map<String, dynamic>;
-
-    if (msg["op"] == OPCodes.dispatch && manager._ws._client._options.ignoredEvents.contains(msg["t"] as String)) {
+    if(rawData["jsonData"] == null) {
       return;
     }
 
-    if (msg["s"] != null) {
-      this._sequence = msg["s"] as int;
+    final discordPayload = rawData["jsonData"] as Map<String, dynamic>;
+
+    if (discordPayload["op"] == OPCodes.dispatch && manager._ws._client._options.ignoredEvents.contains(discordPayload["t"] as String)) {
+      return;
+    }
+
+    if (discordPayload["s"] != null) {
+      this._sequence = discordPayload["s"] as int;
     }
     
-    await _dispatch(msg);
+    await _dispatch(discordPayload);
   }
   
   Future<void> _dispatch(Map<String, dynamic> rawPayload) async {
     switch (rawPayload["op"] as int) {
       case OPCodes.heartbeatAck:
         this._heartbeatAckReceived = true;
-        this._gatewaylatency = DateTime.now().difference(this._lastHeartbeatSent);
+        this._gatewayLatency = DateTime.now().difference(this._lastHeartbeatSent);
 
         break;
       case OPCodes.hello:
-        if (this._sessionId == null || !resume) {
+        if (this._sessionId == null || !_resume) {
           final identifyMsg = <String, dynamic>{
             "token": manager._ws._client._token,
             "properties": <String, dynamic> {
@@ -284,7 +230,7 @@ class Shard implements Disposable {
           identifyMsg["shard"] = <int>[this.id, manager._numShards];
 
           this.send(OPCodes.identify, identifyMsg);
-        } else if (resume) {
+        } else if (_resume) {
           this.send(OPCodes.resume,
               <String, dynamic>{"token": manager._ws._client._token, "session_id": this._sessionId, "seq": this._sequence});
         }
@@ -313,10 +259,10 @@ class Shard implements Disposable {
             this._sessionId = rawPayload["d"]["session_id"] as String;
             manager._ws._client.self = ClientUser._new(rawPayload["d"]["user"] as Map<String, dynamic>, manager._ws._client);
 
-            this.connected = true;
+            this._connected = true;
             manager._logger.info("Shard ${this.id} ready!");
 
-            if (!resume) {
+            if (!_resume) {
               await manager._ws.propagateReady();
             }
 
@@ -466,88 +412,5 @@ class Shard implements Disposable {
     //await this._receiveStream.firstWhere((element) => (element as Map<String, dynamic>)["cmd"] == "TERMINATE_OK");
 
     this._shardIsolate.kill();
-  }
-}
-
-// Decodes zlib compresses string into string json
-Map<String, dynamic> _decodeBytes(dynamic bytes) {
-  if (bytes is String) {
-    return jsonDecode(bytes) as Map<String, dynamic>;
-  }
-
-  final decoded = zlib.decoder.convert(bytes as List<int>);
-  final rawStr = utf8.decode(decoded);
-  return jsonDecode(rawStr) as Map<String, dynamic>;
-}
-
-/*
-Protocol used to communicate with shard isolate.
-  First message delivered to shardHandler will be init message with gateway uri
-
- * DATA - sent along with data received from websocket
- * DISCONNECTED - sent when shard disconnects
- * ERROR - sent when error occurs
-
- * CONNECT - sent when ws connection is established. additional data can contain if reconnected.
- * SEND - sent along with data to send via websocket
-*/
-Future<void> _shardHandler(SendPort shardPort) async {
-  /// Port init
-  final receivePort = ReceivePort();
-  final receiveStream = receivePort.asBroadcastStream();
-
-  final sendPort = receivePort.sendPort;
-  shardPort.send(sendPort);
-
-  /// Initial data init
-  final initData = await receiveStream.first;
-  final gatewayUri = Constants.gatewayUri(initData["gatewayUrl"] as String);
-
-  transport.WebSocket? _socket;
-  StreamSubscription? _socketSubscription;
-
-  transport_vm.configureWTransportForVM();
-
-  // Attempts to connect to ws
-  Future<void> _connect() async {
-    await transport.WebSocket.connect(gatewayUri).then((ws) {
-      shardPort.send({ "cmd" : "CONNECT_ACK" });
-      _socket = ws;
-      _socketSubscription = _socket!.listen((data) {
-        shardPort.send({ "cmd" : "DATA", "jsonData" : _decodeBytes(data) });
-      }, onDone: () async {
-        shardPort.send({ "cmd" : "DISCONNECTED", "errorCode" : _socket!.closeCode, "errorReason" : _socket!.closeReason });
-      }, cancelOnError: true, onError: (err) => shardPort.send({ "cmd" : "ERROR", "error": err.toString(), "errorCode" : _socket!.closeCode, "errorReason" : _socket!.closeReason }));
-    }, onError: (err, __) => shardPort.send({ "cmd" : "ERROR", "error": err.toString(), "errorCode" : _socket!.closeCode, "errorReason" : _socket!.closeReason }));
-  }
-
-  // Connects
-  await _connect();
-
-  await for(final message in receiveStream) {
-    final cmd = message["cmd"];
-
-    if(cmd == "SEND") {
-      if(_socket?.closeCode == null) {
-        _socket?.add(jsonEncode(message["data"]));
-      }
-
-      continue;
-    }
-
-    if(cmd == "CONNECT") {
-      await _socketSubscription?.cancel();
-      await _socket?.close(1000);
-      await _connect();
-
-      continue;
-    }
-/*
-    if(cmd == "TERMINATE") {
-      await _socketSubscription?.cancel();
-      await _socket?.close(1000);
-      shardPort.send({ "cmd" : "TERMINATE_OK" });
-    }
-*/
   }
 }
